🐛 fix: Profile Image
